Fix docker rm call

+++ lib/containers/engine.pm
@@ -324,7 +324,9 @@ sub cleanup_system_host {
 
     # all containers should be stopped before running prune
     # https://github.com/containers/podman/issues/19038
-    $self->_engine_assert_script_run("rm --force --all", 120);
+    if ($self->runtime eq 'podman') {
+        $self->_engine_assert_script_run("rm --force --all", 120);
+    }
     $self->_engine_assert_script_run("system prune -a -f", 300);
 
     if ($assert) {
